Moved elasticsearch client to utils folder and adjusted paths
