+++ test/feed-processor.test.js
@@ -2,7 +2,7 @@ const fixtures = require('./fixtures');
 const processor = require('../src/backend/feed/processor');
 const Feed = require('../src/backend/data/feed');
 
-jest.mock('../src/backend/lib/elastic');
+jest.mock('../src/backend/utils/elastic');
 
 describe('Feed Processor Tests', () => {
   const createFeed = url =>
