+++ test/feeds-json.test.js
@@ -1,7 +1,7 @@
 const request = require('supertest');
 const app = require('../src/backend/web/app');
 
-jest.mock('../src/backend/lib/elastic');
+jest.mock('../src/backend/utils/elastic');
 
 describe('GET "/feed/json" endpoint', () => {
   let response;
