+++ test/feeds.test.js
@@ -4,7 +4,7 @@ const app = require('../src/backend/web/app');
 const Feed = require('../src/backend/data/feed');
 const hash = require('../src/backend/data/hash');
 
-jest.mock('../src/backend/lib/elastic');
+jest.mock('../src/backend/utils/elastic');
 
 describe('test GET /feeds endpoint', () => {
   const createdItems = 150;
